refactor: Cookie 변환 로직을 Cookies 에게 위임
