+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestHeaders.java
@@ -10,8 +10,6 @@ public class RequestHeaders {
     private static final String HEADER_SEPARATOR = System.lineSeparator();
     private static final String HEADER_KEY_VALUE_SPLIT = ":";
     private static final String COOKIE = "Cookie";
-    private static final String COOKIE_SPLITER = ";";
-    private static final String COOKIE_VALUE_SPLITER = "=";
     private final Map<String, String> headers;
 
     public RequestHeaders(String headers) {
@@ -44,25 +42,7 @@ public class RequestHeaders {
         if (!headers.containsKey(COOKIE)) {
             return Optional.empty();
         }
-        return Optional.of(new Cookies(extractCookie(headers.get(COOKIE))));
-    }
-
-    private Map<String, String> extractCookie(String cookies) {
-        return Stream.of(cookies.split(COOKIE_SPLITER))
-            .collect(Collectors.toMap(
-                this::extractKey,
-                this::extractValue
-            ));
-    }
-
-    private String extractKey(String value) {
-        String[] split = value.split(COOKIE_VALUE_SPLITER);
-        return split[0].trim();
-    }
-
-    private String extractValue(String value) {
-        String[] split = value.split(COOKIE_VALUE_SPLITER);
-        return split[1].trim();
+        return Optional.of(new Cookies(headers.get(COOKIE)));
     }
 
     public Map<String, String> getHeaders() {
